+++ packages/perseus-editor/src/widgets/interaction-editor/index.ts
@@ -1,3 +1,1 @@
-import InteractionEditor from "./interaction-editor";
-
-export default InteractionEditor;
+export {default} from "./interaction-editor";